requirements.txt
discord.py
aiosqlite
textblob
groq